Feature: update franchise name endpoint and test

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/bancolombia/r2dbc/entities/FranchiseEntity.java
@@ -15,7 +15,7 @@ public class FranchiseEntity {
 
     @Id
     @Column("franchise_id")
-    private Integer franchiseId;
+    private Long franchiseId;
     @Column("name")
     private String name;
 }
